Reformat all code in the project

+++ src/org/jcodec/containers/mp4/boxes/SampleSizesBox.java
@@ -68,7 +68,7 @@ public class SampleSizesBox extends FullBox {
     @Override
     public void doWrite(ByteBuffer out) {
         super.doWrite(out);
-        out.putInt((int) defaultSize);
+        out.putInt(defaultSize);
 
         if (defaultSize == 0) {
             out.putInt(count);
@@ -77,7 +77,7 @@ public class SampleSizesBox extends FullBox {
                 out.putInt((int) size);
             }
         } else {
-            out.putInt((int)count);
+            out.putInt(count);
         }
     }
